pyproject.toml
[tool.poetry]
name = "bot-for-test"
version = "0.1.0"
description = ""
authors = ["zwickvitaly <zwickvitaly@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiogram = "^3.8.0"
gspread = "^6.1.2"
requests = "^2.32.3"
dateparser = "^1.2.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
